Refactor background tasks into separate file and add error handling
